US270 get timeout from config, config types made int rather than string that gets converted later
Former-commit-id: bc11b679e5050e0dbca655f6df2a8fc4eeb13105

+++ app/services/address_lookup/ordnance_survey/WebServiceImpl.scala
@@ -8,7 +8,7 @@ import services.address_lookup.AddressLookupWebService
 
 final class WebServiceImpl extends AddressLookupWebService {
   private val baseUrl: String = Config.ordnanceSurveyMicroServiceUrl // TODO would it be better to move these to a companion object? And maybe private[this]
-  private val requestTimeout: Int = Config.ordnanceSurveyRequestTimeout.toInt
+  private val requestTimeout: Int = Config.ordnanceSurveyRequestTimeout
 
   def postcodeWithNoSpaces(postcode: String): String = postcode.filter(_ != ' ')
